pyproject.toml
[tool.poetry]
name = "airbnb-clone"
version = "0.1.0"
description = ""
authors = ["yourlucky <yoonchae.na@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
pillow = "^9.4.0"
black = "^23.1.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
